fixed bug in inverse overlapmatrix
There was a problem with a factor of 2 in the bar matrices because they are only half the size
of the ones in the uncoupled version. So some of the factors calculated in the coefficient matrix where half
of the size they should have been.

Also for the proj_bar, this was an error, and this got into the "alpha(i)" function.